requirements.txt
python-dotenv
langchain
langchain-ollama
aiofiles
fastapi
uvicorn
pydantic
requests